Factor out demos and put them into a new directory
